+++ src/compat/pthread/tests/condvar2_1.c
@@ -134,7 +134,7 @@ int pthread_test_condvar2_1()
 
   for (i = 1; i <= NUMTHREADS; i++)
     {
-      assert(pthread_create(&t[i], NULL, mythread, (void *) i) == 0);
+	assert(pthread_create(&t[i], NULL, mythread, (void *) (int64_t) i) == 0);
     }
 
   assert(pthread_mutex_unlock(&mutex) == 0);
