+++ src/compat/pthread/tests/mutex4.c
@@ -57,7 +57,7 @@ static pthread_mutex_t mutex1;
 
 void * unlocker(void * arg)
 {
-  int expectedResult = (int) arg;
+    int expectedResult = (int) (int64_t)arg;
 
   wasHere++;
   assert(pthread_mutex_unlock(&mutex1) == expectedResult);
