+++ src/compat/pthread/tests/semaphore4t.c
@@ -87,7 +87,7 @@ static sem_t s;
 static void *
 thr (void * arg)
 {
-  if ((int) arg == 5)
+  if ((int) (int64_t)  arg == 5)
     {
       // We expect this thread to be cancelled,
       // so sem_wait should return EINTR.
@@ -114,7 +114,7 @@ int pthread_test_semaphore4t(void)
 
   for (i = 1; i <= MAX_COUNT; i++)
     {
-      assert(pthread_create(&t[i], NULL, thr, (void *) i) == 0);
+      assert(pthread_create(&t[i], NULL, thr, (void *)  (int64_t) i) == 0);
       do
         {
           sched_yield();
